+++  my-simple-code/src/mysdk/exconfig/Config.cpp
@@ -137,7 +137,7 @@ bool ConfigFile::SetSource(const char* file, bool ignorecase)
 
 		/* get the length of the file */
 		fseek(f, 0, SEEK_END);
-		length = ftell(f);
+		length = static_cast<int>(ftell(f));
 		buf = new char[length + 1];
 		fseek(f, 0, SEEK_SET);
